T#780: Extract 8 DM routes to src/dm/routes.ts
Phase 2.2 of Backend Modularization (Library #102). Mechanical move,
no logic changes.

**Moved to src/dm/routes.ts (~277L)**:
- 8 DM routes: GET /dm/dashboard, GET /dm/unread-count, POST /dm,
  GET /dm/:name, GET /dm/:name/:other, PATCH /dm/:name/:other/read,
  PATCH /dm/:name/:other/read-all, DELETE /dm/messages/:id

**Cross-domain imports**:
- Role from ../server/rbac.ts
- checkGuestDmRate, checkGuestContentLength, scanForInjection from ../server/guest-safety.ts
- logSecurityEvent from ../server/security-logger.ts
- getDashboard, listConversations, markRead, markAllRead, getMessages-as-getDmMessages from ./handler.ts
- getBeastProfile from ../db/index.ts
- listGuests, getGuestByUsername, getGuestByDisplayName from ../server/guest-accounts.ts

**Server.ts changes**: 275 lines removed. Net: server.ts down ~275 lines (was 5,808, now 5,533).

**Cumulative T#771-T#780**: server.ts -6,430 lines (11,963 → 5,533, -54%)
